% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{calculate_credible_intervals}
\alias{calculate_credible_intervals}
\title{Calculate Credible Intervals}
\usage{
calculate_credible_intervals(predictions, levels)
}
\description{
Calculate Credible Intervals
}
\keyword{internal}
