% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{calculate_test_statistic}
\alias{calculate_test_statistic}
\title{Calculate Test Statistic}
\usage{
calculate_test_statistic(data, stat_name)
}
\arguments{
\item{data}{Numeric vector}

\item{stat_name}{Name of test statistic}
}
\value{
Numeric value
}
\description{
Internal helper to calculate various test statistics
}
\keyword{internal}
