% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_likelihood_guardrails.R
\name{check_negative_correlation}
\alias{check_negative_correlation}
\title{Check for negative correlations between observed and estimated time series}
\usage{
check_negative_correlation(
  obs_cases,
  est_cases,
  obs_deaths,
  est_deaths,
  threshold
)
}
\arguments{
\item{obs_cases}{Matrix of observed cases}

\item{est_cases}{Matrix of estimated cases}

\item{obs_deaths}{Matrix of observed deaths}

\item{est_deaths}{Matrix of estimated deaths}

\item{threshold}{Correlation threshold below which to trigger guardrail}
}
\value{
Character vector of violations found
}
\description{
Check for negative correlations between observed and estimated time series
}
\keyword{internal}
