% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_likelihood_guardrails.R
\name{check_nonsensical_values}
\alias{check_nonsensical_values}
\title{Check for nonsensical values in estimated data}
\usage{
check_nonsensical_values(est_cases, est_deaths, max_cases, max_deaths)
}
\arguments{
\item{est_cases}{Matrix of estimated cases}

\item{est_deaths}{Matrix of estimated deaths}

\item{max_cases}{Maximum reasonable cases per timestep}

\item{max_deaths}{Maximum reasonable deaths per timestep}
}
\value{
Character vector of violations found
}
\description{
Check for nonsensical values in estimated data
}
\keyword{internal}
