% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_country_to_iso.R
\name{convert_country_to_iso}
\alias{convert_country_to_iso}
\title{Convert Country Names to ISO3 or ISO2 Country Codes}
\usage{
convert_country_to_iso(x, iso3 = TRUE)
}
\arguments{
\item{x}{A character vector of country names (e.g., "United States", "UK", "Democratic Republic of Congo").}

\item{iso3}{A logical value. If TRUE (default), returns ISO3 country codes. If FALSE, returns ISO2 country codes.}
}
\value{
A character vector of ISO3 or ISO2 country codes corresponding to the country names.
}
\description{
This function converts country names to their corresponding ISO3 or ISO2 country codes. The function uses the \code{countrycode} package, which allows for multiple spellings and variations of country names.
}
\details{
The function converts country names to ISO3 or ISO2 codes using the \code{countrycode} package. It handles various common and alternative spellings of country names.
}
\examples{
# Convert a single country name to ISO3
convert_country_to_iso("United States")

# Convert a vector of country names to ISO2
convert_country_to_iso(c("United States", "UK", "Democratic Republic of Congo"), iso3 = FALSE)

}
