% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_iso_codes.R
\name{convert_iso_codes}
\alias{convert_iso_codes}
\title{Convert ISO3 to ISO2 or ISO2 to ISO3 Country Codes}
\usage{
convert_iso_codes(iso_codes)
}
\arguments{
\item{iso_codes}{A vector or scalar of ISO2 or ISO3 country codes (e.g., "USA", "GBR", "ZA", "GB").}
}
\value{
A character vector of converted ISO2 or ISO3 country codes corresponding to the provided input.
}
\description{
This function converts a vector or scalar of ISO3 country codes to ISO2 country codes, or vice versa, based on the input. The function uses the \code{countrycode} package and auto-detects the input format.
}
\details{
The function detects whether the input is ISO2 or ISO3 format based on the code length and converts accordingly using the \code{countrycode} package.
}
\examples{
# Convert a vector of ISO3 codes to ISO2 codes
convert_iso_codes(c("USA", "GBR", "DZA"))

# Convert a vector of ISO2 codes to ISO3 codes
convert_iso_codes(c("US", "GB", "DZ"))

}
