% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_iso_to_country.R
\name{convert_iso_to_country}
\alias{convert_iso_to_country}
\title{Convert ISO2 or ISO3 Country Codes to Country Names}
\usage{
convert_iso_to_country(iso)
}
\arguments{
\item{iso}{A vector or scalar of ISO2 or ISO3 country codes (e.g., "USA", "GB", "DZA").}
}
\value{
A character vector of country names corresponding to the provided ISO2 or ISO3 country codes.
}
\description{
This function converts a vector or scalar of ISO2 or ISO3 country codes to their corresponding country names using the \code{countrycode} package.
}
\details{
The function automatically detects if the input is in ISO2 or ISO3 format and returns the corresponding country names. Special handling is applied for Congo and the Democratic Republic of Congo.
}
\examples{
# Convert a vector of ISO3 and ISO2 codes to country names
convert_iso_to_country(c("USA", "GBR", "DZA", "COD"))

}
