% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{create_coverage_plots}
\alias{create_coverage_plots}
\title{Create Coverage Plots}
\usage{
create_coverage_plots(ppc_results, credible_levels, location_names)
}
\description{
Create Coverage Plots
}
\keyword{internal}
