% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{create_density_plots}
\alias{create_density_plots}
\title{Create Density Comparison Plots}
\usage{
create_density_plots(observed, predicted, location_names)
}
\description{
Create Density Comparison Plots
}
\keyword{internal}
