% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create Diagnostic Plots}
\usage{
create_diagnostic_plots(observed, predicted, location_names)
}
\description{
Create Diagnostic Plots
}
\keyword{internal}
