% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{create_ppc_plots}
\alias{create_ppc_plots}
\title{Create PPC Plots}
\usage{
create_ppc_plots(observed_stats, predictive_stats, test_statistics)
}
\arguments{
\item{observed_stats}{List of observed test statistics}

\item{predictive_stats}{List of predictive test statistics}

\item{test_statistics}{Vector of test statistic names}
}
\value{
List of ggplot2 objects
}
\description{
Internal function to create posterior predictive check plots
}
\keyword{internal}
