% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{create_sbc_histogram_plot}
\alias{create_sbc_histogram_plot}
\title{Create SBC Histogram Plot}
\usage{
create_sbc_histogram_plot(sbc_ranks, param_names, p_values)
}
\arguments{
\item{sbc_ranks}{Matrix of SBC ranks}

\item{param_names}{Parameter names}

\item{p_values}{P-values from uniformity tests}
}
\value{
ggplot2 object
}
\description{
Internal function to create SBC diagnostic histograms showing
rank distributions for each parameter.
}
\keyword{internal}
