% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{create_timeseries_plots}
\alias{create_timeseries_plots}
\title{Create Time Series Plots}
\usage{
create_timeseries_plots(
  observed,
  predicted,
  best_model,
  location_names,
  time_labels,
  credible_levels,
  max_traces
)
}
\description{
Create Time Series Plots
}
\keyword{internal}
