% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.check_bounds_vectorized}
\alias{.check_bounds_vectorized}
\title{Check if Samples are Within Bounds}
\usage{
.check_bounds_vectorized(samples, bounds_lower, bounds_upper)
}
\description{
Check if Samples are Within Bounds
}
\keyword{internal}
