% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{.cleanup_npe_resources}
\alias{.cleanup_npe_resources}
\title{Clean up Python and R resources}
\usage{
.cleanup_npe_resources(use_gpu = FALSE, logger = NULL)
}
\description{
Clean up Python and R resources
}
\keyword{internal}
