% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.draw_posterior_samples}
\alias{.draw_posterior_samples}
\title{Draw Posterior Samples (Helper for Rejection Sampling)}
\usage{
.draw_posterior_samples(
  npe_model,
  obs_norm,
  n,
  normalization,
  device,
  return_log_probs,
  torch
)
}
\description{
Draw Posterior Samples (Helper for Rejection Sampling)
}
\keyword{internal}
