% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.extract_bounds_from_model}
\alias{.extract_bounds_from_model}
\title{Extract Parameter Bounds from Model}
\usage{
.extract_bounds_from_model(model)
}
\description{
Extract Parameter Bounds from Model
}
\keyword{internal}
