% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.extract_location_codes}
\alias{.extract_location_codes}
\title{Extract location codes from parameter names}
\usage{
.extract_location_codes(param_names)
}
\description{
Extract location codes from parameter names
}
\keyword{internal}
