% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{.get_distribution_type}
\alias{.get_distribution_type}
\title{Get distribution type for a parameter from priors}
\usage{
.get_distribution_type(param_name, priors)
}
\arguments{
\item{param_name}{Parameter name}

\item{priors}{MOSAIC priors object}
}
\value{
List with distribution info or NULL
}
\description{
Get distribution type for a parameter from priors
}
