% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{.get_npe_config}
\alias{.get_npe_config}
\title{Get NPE configuration with defaults and overrides}
\usage{
.get_npe_config(override = NULL)
}
\description{
Get NPE configuration with defaults and overrides
}
\keyword{internal}
