% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.get_param_description}
\alias{.get_param_description}
\title{Get parameter description from priors.json}
\usage{
.get_param_description(base_name, priors)
}
\description{
Get parameter description from priors.json
}
\keyword{internal}
