% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.is_location_specific}
\alias{.is_location_specific}
\title{Check if parameter is location-specific}
\usage{
.is_location_specific(param_name, location_codes)
}
\description{
Check if parameter is location-specific
}
\keyword{internal}
