% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.parse_location_param}
\alias{.parse_location_param}
\title{Parse location-specific parameter name}
\usage{
.parse_location_param(param_name)
}
\description{
Parse location-specific parameter name
}
\keyword{internal}
