% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ENSO_data.R
\name{.read_enso_forecast_json}
\alias{.read_enso_forecast_json}
\title{Read ENSO Forecast from JSON Configuration}
\usage{
.read_enso_forecast_json()
}
\value{
Data frame with year, month, month_name, variable, value columns
}
\description{
Read ENSO Forecast from JSON Configuration
}
\keyword{internal}
