% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{.reshape_samples}
\alias{.reshape_samples}
\title{Reshape Samples to Matrix (Helper)}
\usage{
.reshape_samples(samples_numpy, n_samples_batch, model)
}
\description{
Reshape Samples to Matrix (Helper)
}
\keyword{internal}
