% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{.validate_train_inputs}
\alias{.validate_train_inputs}
\title{Validate inputs for train_npe function}
\usage{
.validate_train_inputs(
  simulations_file,
  outputs_file,
  output_dir,
  param_names,
  simulation_filter,
  architecture,
  override_spec,
  use_gpu,
  seed,
  verbose,
  logger
)
}
\description{
Validate inputs for train_npe function
}
\keyword{internal}
