% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_africa_shapefile.R
\name{download_africa_shapefile}
\alias{download_africa_shapefile}
\title{Download and Save Africa Shapefile}
\usage{
download_africa_shapefile(PATHS)
}
\arguments{
\item{PATHS}{A list or environment containing the path structure, where \code{PATHS$DATA_PROCESSED} specifies the directory to save the processed shapefile.}
}
\value{
A message indicating that the shapefile has been saved, including the full path to the saved file.
}
\description{
This function downloads the shapefile for African countries using the \code{rnaturalearth} package and saves it in a specified directory. If the output directory does not exist, it will be created.
}
\details{
This function uses the \code{rnaturalearth} package to download the shapefile for the African continent. The shapefile is saved as \code{AFRICA_ADM0.shp} in the directory specified by \code{PATHS$DATA_PROCESSED/shapefiles}. The function will create the directory if it does not already exist.
}
\examples{
\dontrun{
  # Example PATHS object
  PATHS <- list(DATA_PROCESSED = "path/to/processed/data")

  # Download and save the Africa shapefile
  download_africa_shapefile(PATHS)
}

}
