% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effective_aic_range_alternatives.R
\name{effective_aic_alternatives}
\alias{effective_aic_alternatives}
\alias{get_effective_aic_range_chisq}
\alias{get_effective_aic_range_likelihood}
\alias{get_effective_aic_range_empirical}
\title{Alternative Theoretically-Grounded Approaches for Effective AIC Range}
\usage{
get_effective_aic_range_chisq(top_percentile, df = 2)

get_effective_aic_range_likelihood(top_percentile)

get_effective_aic_range_empirical(top_percentile)
}
\arguments{
\item{top_percentile}{Numeric percentage (0-100) of top simulations}

\item{df}{Degrees of freedom for chi-square approach (default: 2)}
}
\description{
These functions demonstrate more principled approaches based on statistical theory
}
\examples{
# Compare different approaches
percentiles <- c(1, 5, 10, 20)

# Current ad hoc formula
current <- sapply(percentiles, function(p) 2 + 2.3 * log10(p))

# Chi-square based
chisq_based <- sapply(percentiles, get_effective_aic_range_chisq)

# Relative likelihood based
likelihood_based <- sapply(percentiles, get_effective_aic_range_likelihood)

}
