% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_symptomatic_prop.R
\name{est_symptomatic_prop}
\alias{est_symptomatic_prop}
\title{Estimate and Visualize Symptomatic Proportion Data}
\usage{
est_symptomatic_prop(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where the symptomatic proportion data and figures should be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_PROCESSED}: Path to the symptomatic proportion data.
\item \strong{DOCS_FIGURES}: Path to save the generated plots.
}}
}
\value{
The function does not return a value but generates and saves plots to a PNG file.
}
\description{
This function reads the symptomatic proportion data, fits a Beta distribution to simulate the proportion of infections that are symptomatic, and generates visualizations comparing estimates from previous studies.
}
\examples{
\dontrun{
# Estimate and visualize symptomatic proportion data
PATHS <- get_paths()
est_symptomatic_prop(PATHS)
}

}
