% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{estimate_overdispersion}
\alias{estimate_overdispersion}
\title{Estimate Overdispersion from Residuals}
\usage{
estimate_overdispersion(observed, predicted, verbose = TRUE)
}
\arguments{
\item{observed}{Observed data}

\item{predicted}{Predicted data}

\item{verbose}{Print messages}
}
\value{
Overdispersion parameter
}
\description{
Estimate Overdispersion from Residuals
}
\keyword{internal}
