% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_posterior.R
\name{evaluate_prior_densities}
\alias{evaluate_prior_densities}
\title{Evaluate Prior Densities for SMC}
\usage{
evaluate_prior_densities(samples, priors, verbose = FALSE)
}
\arguments{
\item{samples}{Matrix of parameter samples}

\item{priors}{List of prior specifications}

\item{verbose}{Print progress}
}
\value{
Vector of log prior densities
}
\description{
Evaluates log p(theta) for each posterior sample using original priors.
Required for SMC importance weight calculation.
}
