% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{extract_observed_data}
\alias{extract_observed_data}
\title{Extract Observed Data from Configuration}
\usage{
extract_observed_data(config_base)
}
\arguments{
\item{config_base}{Configuration object}
}
\value{
List with cases and deaths matrices
}
\description{
Extract Observed Data from Configuration
}
\keyword{internal}
