% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_country_grid_dist.R
\name{generate_country_grid_dist}
\alias{generate_country_grid_dist}
\title{Generate a Grid of Points Within a Country's Shapefile Based on Distance}
\usage{
generate_country_grid_dist(country_shp, distance_km)
}
\arguments{
\item{country_shp}{An \code{sf} object representing the country's shapefile.}

\item{distance_km}{A numeric value specifying the distance (in kilometers) between grid points.}
}
\value{
An \code{sf} object containing the generated grid of points within the country boundary.
}
\description{
This function generates a grid of points within a country's shapefile, where the points are spaced by a specified distance in kilometers.
}
\details{
The function creates a regular grid of points within the polygon defined by the country's shapefile, spaced according to the specified distance in kilometers.
}
\examples{
\dontrun{
# Example usage with distance in kilometers
grid_points_sf <- generate_country_grid_dist(country_shapefile, distance_km = 50)
}
}
