% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{generate_ppc_simulation}
\alias{generate_ppc_simulation}
\title{Generate PPC Samples via Simulation}
\usage{
generate_ppc_simulation(
  best_model_info,
  n_replications,
  config_base,
  PATHS,
  parallel = TRUE,
  n_cores = 4,
  verbose = TRUE
)
}
\arguments{
\item{best_model_info}{Best model information}

\item{n_replications}{Number of replications}

\item{config_base}{Base configuration}

\item{PATHS}{Path information}

\item{parallel}{Use parallel processing}

\item{n_cores}{Number of cores}

\item{verbose}{Print progress}
}
\value{
List with cases and deaths prediction arrays
}
\description{
Generate PPC Samples via Simulation
}
\keyword{internal}
