% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{generate_ppc_statistical}
\alias{generate_ppc_statistical}
\title{Generate PPC Samples via Statistical Resampling}
\usage{
generate_ppc_statistical(
  best_predictions,
  n_replications,
  overdispersion,
  verbose = TRUE
)
}
\arguments{
\item{best_predictions}{Best model predictions}

\item{n_replications}{Number of replications}

\item{overdispersion}{Overdispersion parameter}

\item{verbose}{Print progress}
}
\value{
List with cases and deaths prediction arrays
}
\description{
Generate PPC Samples via Statistical Resampling
}
\keyword{internal}
