% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{generate_predictive_data}
\alias{generate_predictive_data}
\title{Generate Predictive Data Using Forward Model}
\usage{
generate_predictive_data(param_values, metadata, n_samples = 1)
}
\arguments{
\item{param_values}{Vector of parameter values}

\item{metadata}{NPE metadata}

\item{n_samples}{Number of predictive samples to generate}
}
\value{
Matrix of predictive data
}
\description{
Internal helper to generate predictive data given parameters
}
\keyword{internal}
