% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ess_parameter.R
\name{generate_test_data_for_ess}
\alias{generate_test_data_for_ess}
\title{Generate Synthetic Data for Testing calc_model_ess_parameter}
\usage{
generate_test_data_for_ess(
  n_samples = 2000,
  n_params = 5,
  true_ess = c(1500, 1000, 500, 200, 100)
)
}
\arguments{
\item{n_samples}{Number of samples to generate}

\item{n_params}{Number of parameters}

\item{true_ess}{Target effective sample size pattern}
}
\value{
Data frame with synthetic results
}
\description{
Generate Synthetic Data for Testing calc_model_ess_parameter
}
