% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Calculate the Centroid of a Country from its Shapefile}
\usage{
get_centroid(shapefile)
}
\arguments{
\item{shapefile}{A character string representing the file path to a country shapefile (in \code{.shp} format).}
}
\value{
A data frame with the following columns:
\itemize{
\item \strong{iso3}: The ISO3 country code extracted from the shapefile name.
\item \strong{lon}: The longitude of the country's centroid.
\item \strong{lat}: The latitude of the country's centroid.
}
}
\description{
This function reads a country shapefile, calculates its centroid, and returns the centroid coordinates (longitude and latitude) along with the ISO3 country code.
}
\details{
The function reads the shapefile using \strong{\code{sf::st_read()}}, calculates the centroid of the country using \strong{\code{sf::st_centroid()}}, and extracts the coordinates of the centroid with \strong{\code{sf::st_coordinates()}}. The ISO3 code is extracted from the filename of the shapefile, assuming the first three characters of the shapefile name represent the ISO3 country code.
}
\examples{
\dontrun{
# Example usage:
# Assuming you have a shapefile "ZAF_ADM0.shp" for South Africa
shapefile_path <- "path/to/shapefile/ZAF_ADM0.shp"
centroid_data <- get_centroid(shapefile_path)
print(centroid_data)
}
}
