% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_shp.R
\name{get_country_shp}
\alias{get_country_shp}
\title{Get a country shapefile from GeoBoundaries API}
\usage{
get_country_shp(iso3, path_output = NULL)
}
\arguments{
\item{iso3}{A three-letter capitalized character string. Must follow the ISO-3166 Alpha-3 country code
standard (\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}).}

\item{path_output}{A file path of where to save the country shapefile}
}
\value{
sf class shapefile
}
\description{
This function retrieves a shapefile for the given iso3 code from the GeoBoundaries data API.
}
\examples{
\dontrun{

tmp <- get_country_shp(iso3 = 'MCO')
plot(tmp)

}
}
