% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DMI_historical.R
\name{get_DMI_historical}
\alias{get_DMI_historical}
\title{Download and Process Historical DMI Data}
\usage{
get_DMI_historical()
}
\value{
A data frame with columns for year, month, month_name, variable (DMI), and value.
}
\description{
This function downloads the historical Dipole Mode Index (DMI) data from NOAA and processes it into a data frame with columns for year, month (1-12), month_name, variable (DMI), and value.
}
\details{
The DMI data is downloaded from NOAA's historical DMI data page: \url{https://psl.noaa.gov/gcos_wgsp/Timeseries/Data/dmi.had.long.data}. The data includes DMI values representing the difference in sea surface temperature anomalies between the western and eastern Indian Ocean.
}
\examples{
\dontrun{
# Get the historical DMI data
dmi_historical <- get_DMI_historical()

# Display the historical DMI data
print(dmi_historical)
}
}
