% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get Mean or Median Elevation from Downloaded DEM Data Using Country Boundaries}
\usage{
get_elevation(PATHS, iso_codes, method = "median")
}
\arguments{
\item{PATHS}{A list containing paths where DEM data is stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_DEM}: Path to the directory where DEM rasters are saved (GeoTIFF format).
\item \strong{DATA_SHAPEFILES}: Path to the directory where shapefiles are saved.
}}

\item{iso_codes}{A character vector of ISO3 country codes for which the elevation data will be calculated.}

\item{method}{A character string specifying the method to use for calculating elevation statistics. Can be \code{"mean"} or \code{"median"} (default is \code{"median"}).}
}
\value{
A data frame with ISO3 codes, country names, and the calculated mean or median elevation for each country.
}
\description{
This function calculates the mean or median elevation for a given set of ISO3 country codes by loading the downloaded DEM files and extracting raster values within the country's boundary.
}
\examples{
\dontrun{
# Define the ISO3 country codes
iso_codes <- c("ZAF", "KEN", "NGA")

# Get paths for data storage
PATHS <- get_paths()

# Calculate mean elevation for the countries
mean_elevations <- get_elevation(iso_codes, PATHS, method = "mean")
print(mean_elevations)
}
}
