% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_python_paths.R
\name{get_python_paths}
\alias{get_python_paths}
\title{Get Python Environment Paths}
\usage{
get_python_paths()
}
\value{
A list with the following named elements:
\describe{
\item{env}{The full absolute path to the desired Python environment directory
(the root folder of the virtual environment).}
\item{exec}{The expected path to the Python executable within that environment,
constructed using OS-specific conventions.}
\item{norm}{The normalized (canonical) absolute path of the Python executable with
symbolic links and/or relative components resolved.}
}
}
\description{
Returns the paths related to the desired MOSAIC Python environment
}
