% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccine_effectiveness_data.R
\name{get_vaccine_effectiveness_data}
\alias{get_vaccine_effectiveness_data}
\title{Get Vaccine Effectiveness Data}
\usage{
get_vaccine_effectiveness_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the data will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_VACCINE_EFFECTIVENESS}: Path to the directory where the vaccine effectiveness data will be saved.
}}
}
\description{
This function saves a dataset containing vaccine effectiveness data extracted from multiple studies. The data is saved as a CSV file.
}
