% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{identify_best_model}
\alias{identify_best_model}
\title{Identify Best Model from Results}
\usage{
identify_best_model(results, col_ll, param_cols, col_seed, verbose = TRUE)
}
\arguments{
\item{results}{Calibration results data frame}

\item{col_ll}{Column index for likelihood}

\item{param_cols}{Column indices for parameters}

\item{col_seed}{Column index for seed (can be NULL)}

\item{verbose}{Print messages}
}
\value{
List with best model information
}
\description{
Identify Best Model from Results
}
\keyword{internal}
