% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies2.R
\name{install_dependencies2}
\alias{install_dependencies2}
\title{Install R and Python Dependencies for MOSAIC using a Conda Environment at a Fixed Location}
\usage{
install_dependencies2(force = FALSE)
}
\arguments{
\item{force}{Logical. If TRUE, deletes and recreates the conda environment from scratch. Default is FALSE.}
}
\value{
No return value. Side effect: installs R and Python dependencies in a conda environment.
}
\description{
Sets up a conda environment for MOSAIC using the \code{environment.yml} file located in the package.
The environment is stored in a fixed, cross-platform directory ("~/.MOSAIC_conda_env"). Installs
the R packages \code{keras3} and \code{tensorflow} if needed, and ensures that the Keras + TensorFlow
Python backend is available for use in R.
}
