% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa.R
\docType{data}
\name{iso_codes_africa}
\alias{iso_codes_africa}
\title{ISO3 Country Codes for All African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa
}
\description{
A character vector containing the ISO3 country codes for all 54 countries in Africa.
}
\examples{
# Access ISO3 codes for all African countries
iso_codes_africa

}
\keyword{datasets}
