% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa_central.R
\docType{data}
\name{iso_codes_africa_central}
\alias{iso_codes_africa_central}
\title{ISO3 Country Codes for Central African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa_central
}
\description{
A character vector containing the ISO3 country codes for 8 countries in Central Africa.
}
\examples{
# Access ISO3 codes for Central African countries
iso_codes_africa_central

}
\keyword{datasets}
