% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa_east.R
\docType{data}
\name{iso_codes_africa_east}
\alias{iso_codes_africa_east}
\title{ISO3 Country Codes for East African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa_east
}
\description{
A character vector containing the ISO3 country codes for 14 countries in East Africa.
}
\examples{
# Access ISO3 codes for East African countries
iso_codes_africa_east

}
\keyword{datasets}
