% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa_north.R
\docType{data}
\name{iso_codes_africa_north}
\alias{iso_codes_africa_north}
\title{ISO3 Country Codes for North African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa_north
}
\description{
A character vector containing the ISO3 country codes for 6 countries in North Africa.
}
\examples{
# Access ISO3 codes for North African countries
iso_codes_africa_north

}
\keyword{datasets}
