% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa_south.R
\docType{data}
\name{iso_codes_africa_south}
\alias{iso_codes_africa_south}
\title{ISO3 Country Codes for Southern African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa_south
}
\description{
A character vector containing the ISO3 country codes for 7 countries in Southern Africa.
}
\examples{
# Access ISO3 codes for Southern African countries
iso_codes_africa_south

}
\keyword{datasets}
