% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_africa_west.R
\docType{data}
\name{iso_codes_africa_west}
\alias{iso_codes_africa_west}
\title{ISO3 Country Codes for West African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_africa_west
}
\description{
A character vector containing the ISO3 country codes for 16 countries in West Africa.
}
\examples{
# Access ISO3 codes for West African countries
iso_codes_africa_west

}
\keyword{datasets}
