% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_mosaic.R
\docType{data}
\name{iso_codes_mosaic}
\alias{iso_codes_mosaic}
\title{ISO3 Country Codes for MOSAIC Modeling Framework}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_mosaic
}
\description{
A character vector containing the ISO3 country codes for the 40 countries modeled in the MOSAIC framework.
}
\examples{
# Access ISO3 codes for countries modeled in the MOSAIC framework
iso_codes_mosaic

}
\keyword{datasets}
