% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_ssa.R
\docType{data}
\name{iso_codes_ssa}
\alias{iso_codes_ssa}
\title{ISO3 Country Codes for Sub-Saharan African Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_ssa
}
\description{
A character vector containing the ISO3 country codes for 48 countries in Sub-Saharan Africa.
}
\examples{
# Access ISO3 codes for Sub-Saharan African countries
iso_codes_ssa

}
\keyword{datasets}
