% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install_packages.R
\name{load_or_install_packages}
\alias{load_or_install_packages}
\title{Load or Install Required Packages (with GitHub handling for mobility, propvacc, and MOSAIC)}
\usage{
load_or_install_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names to load or install.}
}
\description{
This function checks whether the specified packages are installed. If a package is not installed, it will be installed. Special handling is included for \code{mobility}, \code{propvacc}, and \code{MOSAIC}, which are installed from GitHub.
}
\details{
For most packages, this function installs them from CRAN if they are not already installed. Special cases are handled for the following packages:
\itemize{
\item \strong{mobility}: Installed from GitHub at \url{https://github.com/COVID-19-Mobility-Data-Network/mobility}.
\item \strong{propvacc}: Installed from GitHub at \url{https://github.com/gilesjohnr/propvacc}.
\item \strong{MOSAIC}: Installed from GitHub at \url{https://github.com/InstituteforDiseaseModeling/MOSAIC-pkg}.
}
}
\examples{
\dontrun{
load_or_install_packages(c("ggplot2", "mobility", "propvacc", "MOSAIC"))
}

}
