% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lagged_data.R
\name{make_lagged_data}
\alias{make_lagged_data}
\title{Create Lagged Versions of Data}
\usage{
make_lagged_data(x, lags)
}
\arguments{
\item{x}{A data frame or tibble containing the variables for which lagged versions are to be created.}

\item{lags}{A vector of lag values indicating how many time steps to lag the data. For example, \code{lags = 1:3}
will create 1-lagged, 2-lagged, and 3-lagged versions of each variable.}
}
\value{
A data frame containing the original data and the lagged versions of the variables. The lagged columns
are named in the format \code{<column_name>_lag_<lag>}.
}
\description{
This function generates lagged versions of all columns in a given data frame.
It creates new columns for each lag, allowing time series data to be structured
in a way that includes previous time steps as predictors.
}
\details{
This function takes a data frame and creates new columns that are lagged versions of the original columns.
The number of lags is specified by the \code{lags} argument. The function creates new columns for each lag value,
appending the lag number to the original column names.
}
\examples{
\dontrun{
# Example usage of create_lagged_data:
data <- data.frame(var1 = 1:10, var2 = 11:20)
make_lagged_data(data, lags = 1:2)
}

}
