% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_param_df.R
\name{make_param_df}
\alias{make_param_df}
\title{Get Template for Parameter Values}
\usage{
make_param_df(
  variable_name = NULL,
  variable_description = NULL,
  parameter_distribution = NULL,
  i = NULL,
  j = NULL,
  t = NULL,
  parameter_name = NULL,
  parameter_value = NULL
)
}
\arguments{
\item{variable_name}{A character string representing the name of the variable (e.g., 'sigma').}

\item{variable_description}{A character string describing the variable (e.g., 'proportion symptomatic').}

\item{parameter_distribution}{A character string representing the distribution of the parameter (e.g., 'beta').}

\item{i}{A vector representing the origin locations.}

\item{j}{A vector representing the destination locations.}

\item{t}{A vector representing the time points.}

\item{parameter_name}{A character vector of parameter names (e.g., shape1, shape2 for a beta distribution).}

\item{parameter_value}{A numeric vector of parameter values (corresponding to the parameter names).}
}
\value{
A data frame containing the long-form template for the parameter values. If all arguments are \code{NULL}, an empty data frame is returned.
}
\description{
This function generates a template long-form data frame for parameter values, including origin (\code{i}), destination (\code{j}), and time (\code{t}), used to simulate variables such as sigma.
}
\examples{
# Example usage for sigma
prm <- list(shape1 = 2, shape2 = 5)
param_df <- make_param_df("sigma", "proportion symptomatic", "beta",
                          i = "A", j = "B", t = 1, names(prm), unlist(prm))
print(param_df)

}
