% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_africa_map.R
\name{plot_africa_map}
\alias{plot_africa_map}
\title{Plot Africa Map with Cholera Outbreak Countries}
\usage{
plot_africa_map(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the figure will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DOCS_FIGURES}: Path to the directory where the map image will be saved.
}}
}
\value{
A ggplot object showing the map.
}
\description{
This function creates a map of Africa highlighting countries in Sub-Saharan Africa and those with cholera outbreaks in the past 5 and 10 years. The map is saved as a PNG file in the specified \code{PATHS$DOCS_FIGURES} directory.
}
\details{
The function generates a map of Africa, highlighting Sub-Saharan African countries and those with cholera outbreaks in the past 5 and 10 years. The map is saved as a PNG file.
}
\examples{
\dontrun{
# Assuming PATHS is generated from get_paths()
PATHS <- get_paths()
plot_africa_map(PATHS)
}

}
