% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_data.R
\name{plot_climate_data}
\alias{plot_climate_data}
\title{Plot Weekly Climate Data for a Given Country and Save to File}
\usage{
plot_climate_data(PATHS, country_iso_code)
}
\arguments{
\item{PATHS}{A list containing paths where climate data is stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_CLIMATE}: Path to the directory where the processed climate data is stored.
\item \strong{DOCS_FIGURES}: Path to the directory where the output figures will be saved.
}}

\item{country_iso_code}{A character string representing the ISO3 code of the country for which to plot the climate data.}
}
\value{
The function does not return a value but generates and saves a faceted plot of the weekly climate data for the specified country.
}
\description{
This function loads weekly climate data from a Parquet file for a specified country ISO code and plots a faceted plot showing all available climate variables over time. A vertical line indicates the present day, and the plot is saved to the \code{DOCS_FIGURES} directory.
}
