% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effective_aic_range.R
\name{plot_effective_range}
\alias{plot_effective_range}
\title{Plot Effective Range Function}
\usage{
plot_effective_range(percentiles = seq(0.5, 30, by = 0.5))
}
\arguments{
\item{percentiles}{Numeric vector of percentiles to evaluate (default: 0.5 to 30)}
}
\value{
A plot showing the effective AIC range as a function of percentile
}
\description{
Visualizes the relationship between top percentile and effective AIC range
}
\examples{
\dontrun{
plot_effective_range()
plot_effective_range(percentiles = seq(1, 20, by = 1))
}

}
