% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generation_time.R
\name{plot_generation_time}
\alias{plot_generation_time}
\title{Plot Generation Time as Gamma Distribution for Days and Weeks}
\usage{
plot_generation_time(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the plot will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DOCS_FIGURES}: Path to the directory where the generation time plot will be saved.
}}
}
\description{
This function generates a combined plot for the generation time probability distribution over days (based on a gamma distribution) and aggregated over weeks. The resulting plot is saved as a PNG file.
}
