% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mosaic_country_map.R
\name{plot_mosaic_country_map}
\alias{plot_mosaic_country_map}
\title{Plot MOSAIC Country Map}
\usage{
plot_mosaic_country_map(PATHS)
}
\arguments{
\item{PATHS}{A named list containing file paths required for output. It must include:
\describe{
\item{\strong{DOCS_FIGURES}}{Path to the directory where the map image will be saved.}
}}
}
\value{
Message
}
\description{
This function creates a map of Africa highlighting countries included in the
MOSAIC framework. The map is saved as a PNG file in the directory specified by \code{PATHS$DOCS_FIGURES}.
}
\examples{
\dontrun{

  PATHS <- list(DOCS_FIGURES = "path/to/figures")
  # Generate and save the MOSAIC country map
  plot_mosaic_country_map(PATHS)

}

}
