% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_suspected_cases.R
\name{plot_suspected_cases}
\alias{plot_suspected_cases}
\title{Plot Chi PPV Prior Distributions}
\usage{
plot_suspected_cases(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where the parameter data and figures will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{MODEL_INPUT}: Path to the directory where the parameter data frame is stored.
\item \strong{DOCS_FIGURES}: Path to the directory where the plot will be saved.
}}
}
\description{
This function loads the parameter data frame for the positive predictive value (PPV)
of the cholera clinical case definition from the \code{param_chi_suspected_cases.csv} file
and generates a plot showing the Beta prior distributions for chi_endemic (low/all-settings
estimate) and chi_epidemic (high/outbreak estimate).
}
