% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WHO_vaccination_data.R
\name{plot_WHO_vaccination_data}
\alias{plot_WHO_vaccination_data}
\title{Plot WHO Vaccination Data}
\usage{
plot_WHO_vaccination_data(PATHS)
}
\arguments{
\item{PATHS}{A list of file paths, which should include the following elements:
\describe{
\item{MODEL_INPUT}{The path to the folder containing the processed vaccination data (\code{data_vaccinations.csv}).}
\item{DOCS_FIGURES}{The path to the folder where the plot images will be saved.}
}}
}
\value{
This function returns no R object but saves two PNG files:
\itemize{
\item A faceted plot for vaccination data across all countries saved as \code{"vaccination_by_country.png"}.
\item A zoomed-in plot for a selected country saved as \code{"vaccination_example_<iso_code>.png"}.
}
}
\description{
This function creates plots based on WHO vaccination request data, displaying both raw and redistributed doses, and highlighting the proportion of the population vaccinated in each country.
It produces two types of plots: (1) a faceted plot showing the cumulative vaccination data for all countries, and (2) a detailed plot zooming in on a specific country for a selected date range.
}
\details{
The function performs the following steps:
\itemize{
\item Loads the processed vaccination data from a CSV file.
\item Extracts the last cumulative value for each country, displaying the maximum proportion of the population vaccinated.
\item Creates a faceted plot showing raw doses shipped and redistributed doses over time, for each country, with independent y-axes.
\item Saves the faceted plot as a PNG file in the specified folder.
\item Creates a zoomed-in plot for a specific country and date range, adding text labels to highlight important milestones.
\item Saves the zoomed-in plot as a PNG file.
}
}
\examples{
\dontrun{
PATHS <- list(
  MODEL_INPUT = "path/to/input/data",
  DOCS_FIGURES = "path/to/save/figures"
)
plot_WHO_vaccination_data(PATHS)
}

}
