% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{prepare_data_for_statistics}
\alias{prepare_data_for_statistics}
\title{Prepare Data for Test Statistics Calculation}
\usage{
prepare_data_for_statistics(data)
}
\arguments{
\item{data}{Data frame or vector}
}
\value{
Numeric vector
}
\description{
Internal helper to convert data to vector for statistics
}
\keyword{internal}
