% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_funcs.R
\name{prepare_data_for_statistics_from_matrix}
\alias{prepare_data_for_statistics_from_matrix}
\title{Prepare Data from Matrix Row for Test Statistics}
\usage{
prepare_data_for_statistics_from_matrix(matrix_row)
}
\arguments{
\item{matrix_row}{Numeric vector from matrix row}
}
\value{
Numeric vector
}
\description{
Internal helper to convert matrix row to vector
}
\keyword{internal}
