% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_initial_V1_V2.R
\name{print.mosaic_initial_conditions}
\alias{print.mosaic_initial_conditions}
\title{Print method for mosaic_initial_conditions}
\usage{
\method{print}{mosaic_initial_conditions}(x, ...)
}
\description{
Print method for mosaic_initial_conditions
}
