% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{print.mosaic_ppc}
\alias{print.mosaic_ppc}
\title{Print Method for PPC Results}
\usage{
\method{print}{mosaic_ppc}(x, ...)
}
\arguments{
\item{x}{mosaic_ppc object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for PPC Results
}
