% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{print.mosaic_ppc_diagnostics}
\alias{print.mosaic_ppc_diagnostics}
\title{Print Method for PPC Diagnostics}
\usage{
\method{print}{mosaic_ppc_diagnostics}(x, ...)
}
\arguments{
\item{x}{mosaic_ppc_diagnostics object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for PPC Diagnostics
}
