% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_priors.R
\name{print.mosaic_priors}
\alias{print.mosaic_priors}
\title{Print summary of location priors}
\usage{
\method{print}{mosaic_priors}(x, ...)
}
\arguments{
\item{x}{A priors object returned by get_location_priors}

\item{...}{Additional arguments (ignored)}
}
\description{
Print summary of location priors
}
