% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_climate_data.R
\name{process_climate_data}
\alias{process_climate_data}
\title{Process and Aggregate Climate Data for Each Country}
\usage{
process_climate_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where climate data is stored and processed. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_RAW}: Path to the directory where the raw climate data is stored.
\item \strong{DATA_CLIMATE}: Path to the directory where the processed climate data will be saved.
}}
}
\value{
The function does not return a value. It processes the climate data and saves the daily, weekly, and monthly aggregated datasets for each country in Parquet format.
}
\description{
This function processes the climate data downloaded using the \code{download_climate_data} function. It creates three aggregated datasets: daily, weekly, and monthly, for each country and saves them as Parquet files with the country ISO code included in the file name.
}
