% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_OAG_data.R
\name{process_OAG_data}
\alias{process_OAG_data}
\title{Process and Save OAG Flight Data for Africa}
\usage{
process_OAG_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where raw and processed data are stored.
PATHS is typically the output of the \code{get_paths()} function and should include the following components:
\itemize{
\item \strong{DATA_RAW}: Path to the directory containing the raw data files.
\item \strong{DATA_PROCESSED}: Path to the directory where processed data should be saved.
\item \strong{DATA_SHAPEFILES}: Path to the directory containing shapefiles for African countries.
}}
}
\value{
The function processes the OAG flight data and saves the
aggregated results to the specified processed data paths.
}
\description{
This function processes raw OAG flight data for African countries, including data cleaning,
location and date conversion, and aggregation. The processed data is saved as CSV files for
all African countries and a subset of countries used in the MOSAIC modeling framework.
}
\examples{
\dontrun{
# Define paths for raw and processed data using get_paths()
PATHS <- get_paths()

# Process the OAG data and save the results
process_OAG_data(PATHS)
}
}
