% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz_from_ci.R
\name{qgompertz}
\alias{qgompertz}
\title{Gompertz Distribution Quantile Function}
\usage{
qgompertz(p, b, eta, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Numeric vector. Probabilities in [0,1].}

\item{b}{Numeric. Shape parameter (b > 0).}

\item{eta}{Numeric. Rate parameter (eta > 0).}

\item{lower.tail}{Logical. If TRUE, probabilities are P(X <= x).}

\item{log.p}{Logical. If TRUE, probabilities are given as log(p).}
}
\value{
Numeric vector of quantiles.
}
\description{
Compute the quantile function (inverse CDF) of a Gompertz distribution.
}
