% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf5_to_list.R
\name{read_hdf5_to_list}
\alias{read_hdf5_to_list}
\title{Read an HDF5 File into an R List}
\usage{
read_hdf5_to_list(file_path)
}
\arguments{
\item{file_path}{Path to the .h5 or .h5.gz file.}
}
\value{
A named R list representing the hierarchical structure of the HDF5 file.
}
\description{
Reads an HDF5 file (supports .h5 or .h5.gz formats) into a named R list.
}
