% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_MOSAIC_python_env.R
\name{remove_python_env}
\alias{remove_python_env}
\title{Remove the MOSAIC Python Environment}
\usage{
remove_python_env(force = FALSE)
}
\arguments{
\item{force}{Logical. If TRUE, removes the environment without prompting for confirmation. Default is FALSE.}
}
\value{
No return value.
}
\description{
This function checks whether the MOSAIC Python virtual environment exists at "~/.virtualenvs/r-mosaic".
If it exists, it removes the directory and resets the RETICULATE_PYTHON global variable.
}
