% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{reorganize_plots_by_location}
\alias{reorganize_plots_by_location}
\title{Reorganize Plots by Location}
\usage{
reorganize_plots_by_location(all_plots, location_names)
}
\description{
Reorganize Plots by Location
}
\keyword{internal}
