% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_ppc.R
\name{run_best_model_once}
\alias{run_best_model_once}
\title{Run Best Model Once}
\usage{
run_best_model_once(best_model_info, config_base, PATHS)
}
\arguments{
\item{best_model_info}{Best model information}

\item{config_base}{Base configuration}

\item{PATHS}{Path information}
}
\value{
List with cases and deaths predictions
}
\description{
Run Best Model Once
}
\keyword{internal}
