% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ppc.R
\name{save_plot_list}
\alias{save_plot_list}
\title{Save List of Plots}
\usage{
save_plot_list(plots, output_dir, prefix, format, width, height, verbose)
}
\description{
Save List of Plots
}
\keyword{internal}
