% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_colors.R
\name{scale_color_mosaic_d}
\alias{scale_color_mosaic_d}
\alias{scale_colour_mosaic_d}
\alias{scale_fill_mosaic_d}
\title{MOSAIC Discrete Color/Fill Scales for ggplot2}
\usage{
scale_color_mosaic_d(palette = "default", direction = 1, alpha = 1, ...)

scale_colour_mosaic_d(palette = "default", direction = 1, alpha = 1, ...)

scale_fill_mosaic_d(palette = "default", direction = 1, alpha = 1, ...)
}
\arguments{
\item{palette}{Palette name: \code{"default"} or \code{"extended"}.}

\item{direction}{\code{1} or \code{-1}.}

\item{alpha}{Transparency (0-1).}

\item{...}{Passed to \code{\link[ggplot2]{discrete_scale}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Apply the MOSAIC qualitative palette to discrete aesthetics.
}
\examples{
\dontrun{
ggplot(df, aes(x, y, color = group)) + geom_point() + scale_color_mosaic_d()
}

}
