% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_colors.R
\name{scale_color_mosaic_div}
\alias{scale_color_mosaic_div}
\alias{scale_colour_mosaic_div}
\alias{scale_fill_mosaic_div}
\title{MOSAIC Diverging Color/Fill Scales for ggplot2}
\usage{
scale_color_mosaic_div(palette = "blue_red", direction = 1, ...)

scale_colour_mosaic_div(palette = "blue_red", direction = 1, ...)

scale_fill_mosaic_div(palette = "blue_red", direction = 1, ...)
}
\arguments{
\item{palette}{Palette name: \code{"blue_red"}, \code{"blue_orange"},
\code{"green_purple"}.}

\item{direction}{\code{1} or \code{-1}.}

\item{...}{Passed to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Apply a MOSAIC diverging palette to continuous aesthetics.
}
