% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_openmeteo_api_key.R
\name{set_openmeteo_api_key}
\alias{set_openmeteo_api_key}
\title{Set OpenMeteo API Key}
\usage{
set_openmeteo_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{A character string representing the OpenMeteo API key. If \code{NULL}, the function will prompt for the key interactively.}
}
\value{
A message confirming that the API key has been set, and instructions for retrieving the key.
}
\description{
This function sets the OpenMeteo API key and stores it in a global option called \code{openmeteo_api_key}.
If an API key is provided as an argument, it will be used. If not, the function will prompt for the key interactively.
}
\details{
The API key is stored in a global option called \code{openmeteo_api_key} using \code{options()}.
The key can be retrieved at any time using \code{getOption("openmeteo_api_key")}.
}
\examples{
# Set the OpenMeteo API key interactively
set_openmeteo_api_key()

# Set the OpenMeteo API key programmatically
set_openmeteo_api_key("your-api-key-here")

# Access the API key
getOption("openmeteo_api_key")

}
