% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python_env.R
\name{check_python_env}
\alias{check_python_env}
\title{Check Python Environment for MOSAIC}
\usage{
check_python_env()
}
\value{
No return value. Outputs messages to the console.
}
\description{
Checks whether the current R session is attached to the MOSAIC Python environment.
The environment is fixed to "~/.virtualenvs/r-mosaic". If the environment is missing,
it instructs the user to run MOSAIC::install_dependencies(). If a different Python interpreter
is attached, it alerts the user with a restart suggestion.
}
