% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz_from_ci.R
\name{dgompertz}
\alias{dgompertz}
\title{Gompertz Distribution Density Function}
\usage{
dgompertz(x, b, eta, log = FALSE)
}
\arguments{
\item{x}{Numeric vector. Values at which to evaluate the density.}

\item{b}{Numeric. Shape parameter (b > 0).}

\item{eta}{Numeric. Rate parameter (eta > 0).}

\item{log}{Logical. If TRUE, return log density.}
}
\value{
Numeric vector of density values.
}
\description{
Compute the probability density function of a Gompertz distribution.
}
\examples{
x <- seq(0, 5, length.out = 100)
y <- dgompertz(x, b = 2, eta = 3)
plot(x, y, type = "l", main = "Gompertz PDF")

}
