% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate_data.R
\name{download_climate_data}
\alias{download_climate_data}
\title{Download and Save Climate Data for Multiple Countries (Parquet Format, Single Model)}
\usage{
download_climate_data(
  PATHS,
  iso_codes,
  n_points,
  date_start,
  date_stop,
  climate_model,
  api_key
)
}
\arguments{
\item{PATHS}{A list containing paths where raw and processed data are stored.
PATHS is typically the output of the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_SHAPEFILES}: Path to the directory containing country shapefiles.
\item \strong{DATA_CLIMATE}: Path to the directory where processed climate data will be saved.
}}

\item{iso_codes}{A character vector of ISO3 country codes for which climate data should be downloaded.}

\item{n_points}{An integer specifying the number of grid points to generate within each country for which climate data will be downloaded.}

\item{date_start}{A character string representing the start date for the climate data (in "YYYY-MM-DD" format).}

\item{date_stop}{A character string representing the end date for the climate data (in "YYYY-MM-DD" format).}

\item{climate_model}{A single character string representing the climate model to use. Available models include:
\itemize{
\item \strong{CMCC_CM2_VHR4}
\item \strong{FGOALS_f3_H}
\item \strong{HiRAM_SIT_HR}
\item \strong{MRI_AGCM3_2_S}
\item \strong{EC_Earth3P_HR}
\item \strong{MPI_ESM1_2_XR}
\item \strong{NICAM16_8S}
}}

\item{api_key}{A character string representing the API key required to access the climate data API.}
}
\value{
The function does not return a value. It downloads the climate data for each country and saves the results as Parquet files in the specified directory.
}
\description{
This function downloads daily climate data for a list of specified countries, saving the data as Parquet files. The data includes both historical and future climate variables at grid points within each country for a specified climate model.
}
\details{
This function uses country shapefiles to generate a grid of points within each country, at which climate data is downloaded. The function retrieves climate data for the specified date range (\code{date_start} to \code{date_stop}) and the specified climate model. The data is saved for each country in a Parquet file named \verb{climate_data_\{climate_model\}_\{date_start\}_\{date_stop\}_\{ISO3\}.parquet}.

The climate data variables include temperature, wind speed, cloud cover, precipitation, and more. The function retrieves data from a single climate model.
}
\examples{
\dontrun{
# Define paths for raw and processed data using get_paths()
PATHS <- get_paths()

# ISO3 country codes for African countries
iso_codes <- c("ZAF", "KEN", "NGA")

# API key for climate data API
api_key <- "your-api-key-here"

# Download climate data for a specified model and save it for the specified countries
download_climate_data(PATHS, iso_codes, n_points = 5,
                      date_start = "1970-01-01", date_stop = "2030-12-31",
                      climate_model = "MRI_AGCM3_2_S", api_key)
}

}
