% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_country_DEM.R
\name{download_country_DEM}
\alias{download_country_DEM}
\title{Download 1km DEM Raster for Multiple Countries and Save to Files}
\usage{
download_country_DEM(PATHS, iso_codes, zoom_level = 6)
}
\arguments{
\item{PATHS}{A list containing paths where DEM data will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_DEM}: Path to the directory where DEM rasters will be saved.
}
#' @param iso_codes A character vector of ISO3 country codes for which DEM data should be downloaded.}

\item{zoom_level}{An integer representing the zoom level for the DEM data. Zoom levels 6-8 are recommended for 1km resolution (default = 6).}
}
\value{
The function does not return a value. It downloads the DEM data for each country and saves the results as GeoTIFF files in the specified directory.
}
\description{
This function downloads 1km resolution DEM (Digital Elevation Model) raster data for a given vector of ISO3 country codes, and writes each raster to a GeoTIFF file.
}
\examples{
\dontrun{
# Define the ISO3 country codes
iso_codes <- c("ZAF", "KEN", "NGA")

# Get paths for data storage
PATHS <- get_paths()

# Download DEM rasters for the countries
download_country_DEM(iso_codes, PATHS)
}
}
