% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_epidemic_peaks.R
\name{est_epidemic_peaks}
\alias{est_epidemic_peaks}
\title{Estimate Epidemic Peaks from Cholera Time Series Data}
\usage{
est_epidemic_peaks(PATHS)
}
\arguments{
\item{PATHS}{List of paths generated by get_paths() containing required data locations}
}
\value{
A data frame containing epidemic peak parameters
}
\description{
This function identifies and characterizes epidemic peaks in cholera incidence
time series data for all countries with available data. It detects peak timing,
magnitude, and duration of epidemic periods.
}
\examples{
\dontrun{
# Set up paths and run epidemic peak estimation
set_root_directory("~/MOSAIC")
PATHS <- get_paths()

# Estimate epidemic peaks for all countries
epidemic_results <- est_epidemic_peaks(PATHS)
}

}
