% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_immune_decay_vaccine.R
\name{est_immune_decay_vaccine}
\alias{est_immune_decay_vaccine}
\title{Estimate Vaccine Immune Decay}
\usage{
est_immune_decay_vaccine(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where data and figures are saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{MODEL_INPUT}: Path to the directory where parameter files will be saved.
\item \strong{DOCS_FIGURES}: Path to the directory where the plots will be saved.
}}
}
\description{
This function estimates oral cholera vaccine (OCV) immune decay by fitting exponential decay models
to meta-regression effectiveness data from Xu et al (2024). It estimates decay parameters (omega) and
initial effectiveness (phi) for both one-dose and two-dose regimens, and generates plots for predicted
vaccine effectiveness over time.
}
\examples{
\dontrun{
# Assuming PATHS is generated from get_paths()
PATHS <- get_paths()
est_vaccine_immune_decay(PATHS)
}
}
\references{
Xu et al. (2024). Effectiveness of Oral Cholera Vaccines: A Systematic Review and Meta-Analysis.
medRxiv. https://doi.org/10.1101/2024.08.13.24311930v2
}
