% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mobility.R
\name{est_mobility}
\alias{est_mobility}
\title{Estimate and Visualize Mobility Model for MOSAIC Data}
\usage{
est_mobility(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths to directories for raw and processed data. Typically, \code{PATHS} is the output of \code{get_paths()} and should include:
\itemize{
\item \strong{DATA_OAG}: Path to the directory containing flight data.
\item \strong{DATA_SHAPEFILES}: Path to the directory containing country shapefiles.
\item \strong{DATA_DEMOGRAPHICS}: Path to the directory containing demographic data.
\item \strong{MODEL_INPUT}: Path to the directory where model matrices will be saved.
\item \strong{DOCS_FIGURES}: Path to the directory where figures will be saved.
}}
}
\value{
The function does not return a value. It processes data, fits the mobility model, saves the results as CSV files, and creates several visualizations that are saved as PNG files.
}
\description{
This function processes flight data, population data, and distance data to fit a mobility model using the \code{mobility} package. The model estimates both the probability of travel from each origin country (departure) and the probability of travel from each origin to a destination (diffusion). It then saves the model matrices and results to CSV files and generates plots for visualization.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads flight data and constructs a mobility matrix (\code{M}).
\item Constructs a distance matrix (\code{D}) from country centroids based on shapefiles.
\item Retrieves population sizes (\code{N}) from demographic data.
\item Ensures that dimensions of \code{M}, \code{D}, and \code{N} match.
\item Fits the probability of travel (departure) and diffusion (origin-destination) using the \code{mobility} package.
\item Saves the results, including the mobility matrix (\code{M}), distance matrix (\code{D}), and population vector (\code{N}) as CSV files.
\item Saves several plots related to the mobility model, including a heatmap of the flight data, a network plot of flight data, and plots of travel probabilities and diffusion probabilities.
}
}
\examples{
\dontrun{
# Assuming PATHS is generated from the get_paths function
PATHS <- get_paths()

# Fit the mobility model and generate outputs
est_mobility(PATHS)
}

}
