% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mobility.R
\name{est_mobility}
\alias{est_mobility}
\title{Fit Mobility Model Using Flight Data and Distance Matrices}
\usage{
est_mobility(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where data, models, and figures are stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_OAG}: Path to the directory containing flight data (OAG).
\item \strong{DATA_SHAPEFILES}: Path to the directory containing shapefiles for African countries.
\item \strong{DATA_DEMOGRAPHICS}: Path to the directory containing demographic data.
\item \strong{MODEL_INPUT}: Path to the directory where the model input files will be saved.
\item \strong{DOCS_FIGURES}: Path to the directory where figures will be saved.
}}
}
\value{
The function saves mobility matrices (M, D, N), travel probabilities (tau_j), and diffusion matrices (pi_ij) as CSV files in the specified directory. It also generates and saves visualizations as PNG files.
}
\description{
This function estimates the departure probability (tau_i) and diffusion model (pi_ij) using flight data, distance matrices, and population sizes. It processes flight data, builds mobility and distance matrices, fits the mobility model, and saves the results in CSV format.
}
