% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_seasonal_dynamics.R
\name{est_seasonal_dynamics}
\alias{est_seasonal_dynamics}
\title{Estimate Seasonal Dynamics for Cholera and Precipitation Using Fourier Series}
\usage{
est_seasonal_dynamics(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where raw and processed data are stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_CLIMATE}: Path to the directory where climate data is stored.
\item \strong{DATA_SHAPEFILES}: Path to the directory containing country shapefiles.
\item \strong{DATA_WHO_WEEKLY}: Path to the directory containing cholera data.
\item \strong{MODEL_INPUT}: Path to the directory where model input files will be saved.
}}
}
\value{
The function saves the parameter estimates, fitted values, and processed data to a CSV file.
}
\description{
This function retrieves historical precipitation data, processes cholera case data, and fits seasonal dynamics models using a double Fourier series. The results are saved to a CSV file, including parameter estimates, fitted values, and processed data.
}
