% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_seasonal_dynamics.R
\name{est_seasonal_dynamics}
\alias{est_seasonal_dynamics}
\title{Get Seasonal Dynamics Data for Cholera and Precipitation}
\usage{
est_seasonal_dynamics(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where raw and processed data are stored. Typically generated by the \code{get_paths()} function.}
}
\value{
The function saves the parameter estimates, fitted values, and processed data as CSV files.
}
\description{
This function retrieves historical precipitation data, processes cholera case data, fits seasonal dynamics models using Fourier series,
and saves the results to a CSV file.
}
