% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_vaccine_effectiveness.R
\name{est_vaccine_effectiveness}
\alias{est_vaccine_effectiveness}
\title{Estimate Vaccine Effectiveness Decay Model}
\usage{
est_vaccine_effectiveness(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the parameter data and predictions will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{MODEL_INPUT}: Path to the directory where the parameter estimates and predictions will be saved.
\item \strong{DATA_VACCINE_EFFECTIVENESS}: Path to the directory where vaccine effectiveness data is located.
}}
}
\value{
The function does not return any values but saves the following CSV files:
\itemize{
\item \code{param_vaccine_effectiveness.csv}: Parameter estimates for the initial vaccine effectiveness and decay rate, including beta distribution parameters for uncertainty.
\item \code{pred_vaccine_effectiveness.csv}: Predicted vaccine effectiveness over time (days after vaccination).
}
}
\description{
This function fits a proportional decay model to vaccine effectiveness data and estimates the parameters of the model, saving the results to CSV files. The model assumes that vaccine effectiveness decays over time according to a proportional decay function. Additionally, it fits a beta distribution to the initial vaccine effectiveness to capture the uncertainty in the estimate.
}
\details{
The function uses a non-linear least squares (NLS) approach, specifically the Levenberg-Marquardt algorithm from the \code{nls.lm} function in the \code{minpack.lm} package, to fit the vaccine effectiveness decay model. The model is defined by two parameters:
\itemize{
\item \strong{\eqn{\phi}}: Initial vaccine effectiveness at time zero.
\item \strong{\eqn{\omega}}: Decay rate of vaccine effectiveness over time.
}

The standard errors for the parameter estimates are extracted from the covariance matrix of the fitted model. Confidence intervals for the initial vaccine effectiveness are used to fit a beta distribution to capture uncertainty in the estimate.

After fitting the model, the function generates a prediction of vaccine effectiveness over time and saves the predictions, along with the parameter estimates and beta distribution parameters, to CSV files in the specified \code{MODEL_INPUT} directory.
}
\examples{
\dontrun{
# Assuming PATHS is generated from get_paths()
PATHS <- get_paths()

# Estimate vaccine effectiveness and save results
est_vaccine_effectiveness(PATHS)
}

}
