% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DMI_forecast.R
\name{get_DMI_forecast}
\alias{get_DMI_forecast}
\title{Get DMI (Dipole Mode Index) Forecast Data}
\usage{
get_DMI_forecast()
}
\value{
A data frame with columns for year, month (numeric), month_name, variable (DMI), and value. The DMI is a measure of the difference in sea surface temperature anomalies between the western and eastern Indian Ocean and is used to describe the Indian Ocean Dipole (IOD).
}
\description{
This function retrieves the manually extracted Dipole Mode Index (DMI) forecast data from the Bureau of Meteorology's IOD forecast page.
}
\details{
The IOD forecast is manually extracted from the Bureau of Meteorology's IOD forecast page: \url{http://www.bom.gov.au/climate/ocean/outlooks/#region=IOD}. The data includes DMI values over a series of months and years, representing forecasts of sea surface temperature anomalies. Negative DMI values indicate cooler waters in the west, while positive DMI values indicate warmer waters in the west of the Indian Ocean.
}
\examples{
\dontrun{
# Get the DMI forecast data
dmi_forecast <- get_DMI_forecast()

# Display the DMI forecast data
print(dmi_forecast)
}
}
