% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get Mean Elevation Data for Multiple Countries Using Shapefiles}
\usage{
get_elevation(PATHS, n_points, api_key = NULL)
}
\arguments{
\item{PATHS}{A list containing the paths to where the elevation data should be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_SHAPEFILES}: Path to the directory containing country shapefiles.
\item \strong{DATA_ELEVATION}: Path to the directory where the elevation data will be saved.
}}

\item{n_points}{An integer specifying the number of points to generate within each country's boundary.}

\item{api_key}{A character string representing the API key for the Open-Meteo API.}
}
\value{
A data frame with the mean elevation for each country in \code{iso_codes_mosaic}.
}
\description{
This function retrieves elevation data for a specified number of points within the boundary of each country's shapefile from the \code{iso_codes_mosaic} list using the Open-Meteo API. It returns the mean elevation for each country.
}
\details{
The function loads the shapefiles for all countries in \code{iso_codes_mosaic} from the specified directory, generates \code{n_points} random points within each country's boundary, and queries the Open-Meteo API for the elevation data at each point. It returns the mean elevation for each country and saves the data as a CSV file.
}
\examples{
\dontrun{
# Example usage to get mean elevation data for all countries in iso_codes_mosaic
n_points <- 10
api_key <- "your_api_key_here"
PATHS <- get_paths()
mean_elevations <- get_elevation(n_points, api_key, PATHS)
print(mean_elevations)
}
}
