% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generation_time_distribution.R
\name{get_generation_time_distribution}
\alias{get_generation_time_distribution}
\title{Estimate the Generation Time Distribution for Days and Weeks}
\usage{
get_generation_time_distribution(PATHS, mean_generation_time)
}
\arguments{
\item{PATHS}{A list containing the paths where the output tables will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{MODEL_INPUT}: Path to the directory where model input data will be saved.
\item \strong{DOCS_TABLES}: Path to the directory where day-wise and week-wise generation time tables will be saved.
}}

\item{mean_generation_time}{A numeric value representing the desired mean generation time (in days). This value is used to adjust the shape and rate parameters of the gamma distribution.}
}
\description{
This function generates the probability distribution of the generation time for cholera based on a gamma distribution. It allows the specification of the mean generation time, which adjusts the shape and rate parameters. The day-wise and week-wise probabilities are saved as CSV files.
}
