% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Generate Directory Paths for the MOSAIC Project}
\usage{
get_paths(root = NULL)
}
\arguments{
\item{root}{A string specifying the root directory of the MOSAIC project. All other paths will be generated relative to this root.}
}
\value{
A named list containing the following paths:
\item{ROOT}{The root directory provided by the user.}
\item{DATA_RAW}{Path to the raw data directory (under "MOSAIC-data/data/raw").}
\item{DATA_PROCESSED}{Path to the processed data directory (under "MOSAIC-data/data/processed").}
\item{MODEL_INPUT}{\code{NULL}, reserved for the path to the model input (if applicable).}
\item{MODEL_OUTPUT}{\code{NULL}, reserved for the path to the model output (if applicable).}
\item{DOCS_FIGURES}{Path to the figures directory (under "MOSAIC-docs/figures").}
\item{DOCS_TABLES}{Path to the tables directory (under "MOSAIC-docs/tables").}
\item{DOCS_PARAMS}{Path to the parameters directory (under "MOSAIC-docs/parameters").}
}
\description{
The \code{get_paths()} function generates a structured list of file paths for different data and document directories in the MOSAIC project, based on the provided root directory.
}
\details{
This function helps organize the directory structure for data and document storage in the MOSAIC project by generating paths for raw data, processed data, figures, tables, and parameters. The paths are returned as a named list and can be used to streamline the access to various project-related directories.
}
\examples{
\dontrun{
root_dir <- "/{full file path}/MOSAIC"
PATHS <- get_paths(root_dir)
print(PATHS$DATA_RAW)
}
}
