% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspected_cases.R
\name{get_suspected_cases}
\alias{get_suspected_cases}
\title{Get Proportion of Suspected Cholera Cases and Save Parameter Data Frame}
\usage{
get_suspected_cases(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where the parameter data will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{MODEL_INPUT}: Path to the directory where the parameter data frame will be saved.
}}
}
\value{
A data frame containing the parameter values for the proportion of suspected cholera cases (rho).
}
\description{
This function generates the parameters for the proportion of suspected cholera cases (rho) and saves it as a parameter data frame.
}
