% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install R and Python Dependencies for MOSAIC}
\usage{
install_dependencies(force = FALSE)
}
\arguments{
\item{force}{Logical. If TRUE, deletes and recreates the conda environment from scratch. Default is FALSE.}
}
\value{
No return value
}
\description{
Sets up a conda environment for MOSAIC using the \code{environment.yml} file located in the package.
The environment is stored in a fixed directory ("~/.virtualenvs/r-mosaic") and installs the LASER disease
transmission model simulation tool. This function also installs the R packages \code{keras3} and \code{tensorflow},
and ensures that the Keras + TensorFlow Python backend is available for use in R.
}
