% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_codes_who_afro.R
\docType{data}
\name{iso_codes_who_afro}
\alias{iso_codes_who_afro}
\title{ISO3 Country Codes for WHO AFRO Region Countries}
\format{
A character vector of ISO3 country codes.
}
\usage{
iso_codes_who_afro
}
\description{
A character vector containing the ISO3 country codes for 47 countries in the WHO African Region (AFRO).
}
\examples{
# Access ISO3 codes for WHO AFRO Region countries
iso_codes_who_afro

}
\keyword{datasets}
