% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LASER_config.R
\name{make_LASER_config}
\alias{make_LASER_config}
\title{Create a JSON Configuration File for LASER}
\usage{
make_LASER_config(
  output_file_path = NULL,
  date_start = NULL,
  date_stop = NULL,
  location_id = NULL,
  location_name = NULL,
  N_j_initial = NULL,
  S_j_initial = NULL,
  E_j_initial = NULL,
  I_j_initial = NULL,
  R_j_initial = NULL,
  V1_j_initial = NULL,
  V2_j_initial = NULL,
  b_jt = NULL,
  d_jt = NULL,
  nu_jt = NULL,
  phi_1 = NULL,
  phi_2 = NULL,
  omega_1 = NULL,
  omega_2 = NULL,
  epsilon = NULL,
  gamma_1 = NULL,
  gamma_2 = NULL,
  mu = NULL,
  rho = NULL,
  sigma = NULL,
  beta_j0_hum = NULL,
  beta_j_seasonality = NULL,
  tau_i = NULL,
  pi_ij = NULL,
  alpha_1 = NULL,
  alpha_2 = NULL,
  beta_j0_env = NULL,
  theta_j = NULL,
  psi_jt = NULL,
  zeta_1 = NULL,
  zeta_2 = NULL,
  kappa = NULL,
  delta_min = NULL,
  delta_max = NULL
)
}
\arguments{
\item{output_file_path}{A character string representing the full file path of the output JSON file (e.g., 'path/to/parameters.json'). Must have a .json extension. If \code{NULL}, no file is written.
\subsection{Initialization}{
}}

\item{date_start}{Start date for the simulation period in "YYYY-MM-DD" format. If provided as a character string, it will be converted to a Date object.}

\item{date_stop}{End date for the simulation period in "YYYY-MM-DD" format. If provided as a character string, it will be converted to a Date object.}

\item{location_id}{A vector of integers giving the numerical index for metapopulations.}

\item{location_name}{A character vector giving the names of each metapopulation location. The order and names here must match those used in the initial population vectors.}

\item{N_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the total initial population size for each location. The names must match the values in \code{location_name}.}

\item{S_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of susceptible individuals for each location. Names must match \code{location_name}.}

\item{E_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of exposed individuals for each location. Names must match \code{location_name}.}

\item{I_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of infected individuals for each location. Names must match \code{location_name}.}

\item{R_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of recovered individuals for each location. Names must match \code{location_name}.}

\item{V1_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of individuals in vaccine compartment V1 for each location. Names must match \code{location_name}.}

\item{V2_j_initial}{A named numeric or integer vector of length equal to \code{location_id} that gives the starting number of individuals in vaccine compartment V2 for each location. Names must match \code{location_name}.
\subsection{Demographics}{
}}

\item{b_jt}{A matrix of birth rates with dimensions corresponding to the number of locations (rows) and the number of daily time steps (columns). All values must be numeric and non-negative.}

\item{d_jt}{A matrix of mortality rates with dimensions corresponding to the number of locations (rows) and the number of daily time steps (columns). All values must be numeric and non-negative.
\subsection{Immune Dynamics}{
}}

\item{nu_jt}{A matrix of vaccination rates for each location (\code{j}) and time step (\code{t}). Must have rows equal to \code{length(location_id)} and columns equal to the daily sequence from \code{date_start} to \code{date_stop}.}

\item{phi_1}{Effectiveness of one dose of Oral Cholera Vaccine (OCV). Must be numeric and within the range \link{0, 1}.}

\item{phi_2}{Effectiveness of two doses of Oral Cholera Vaccine (OCV). Must be numeric and within the range \link{0, 1}.}

\item{omega_1}{Waning immunity rate of individuals vaccinated with one dose of OCV. Must be a numeric scalar greater than or equal to zero.}

\item{omega_2}{Waning immunity rate of individuals vaccinated with two doses of OCV. Must be a numeric scalar greater than or equal to zero.}

\item{epsilon}{Waning immunity rate of recovered individuals. Must be a numeric scalar greater than or equal to zero.}

\item{gamma_1}{Recovery rate of infected individuals with severe infection. Must be a numeric scalar greater than or equal to zero.}

\item{gamma_2}{Recovery rate of infected individuals with mild or asymptomatic infection. Must be a numeric scalar greater than or equal to zero.
\subsection{Observation Processes}{
}}

\item{mu}{Mortality rate due to V. cholerae infection. Must be a numeric scalar greater than or equal to zero.}

\item{rho}{Proportion of suspected cholera cases that are true infections. Must be a numeric scalar between 0 and 1.}

\item{sigma}{Proportion of symptomatic V. cholerae infections. Must be a numeric scalar between 0 and 1.
\subsection{Force of Infection (human-to-human)}{
}}

\item{beta_j0_hum}{Baseline human-to-human transmission rate for each location. Must be a numeric vector of length equal to \code{location_id} and values greater than or equal to zero.}

\item{beta_j_seasonality}{The seasonal derivation from mean transmission (beta_j0_hum) for each location. Must be a matrix with rows equal to \code{length(location_id)} and 366 columns representing the annual daily scale.}

\item{tau_i}{Departure probability for each origin location. Must be a numeric vector of length equal to \code{location_id} and values between 0 and 1.}

\item{pi_ij}{A matrix of travel probabilities between origin and destination locations. Must have dimensions equal to \code{length(location_id)} x \code{length(location_id)} and values between 0 and 1.}

\item{alpha_1}{Transmission parameter for the level population mixing between infectious (I) and susceptible (S) individuals. Must be a numeric scalar between 0 and 1.}

\item{alpha_2}{Transmission parameter determining the extent to which transmission is frequency dependent (\eqn{\alpha_2 = 1}) or density dependent (\eqn{\alpha_2 \approx 0}). Must be a numeric scalar between 0 and 1.
\subsection{Force of Infection (environment-to-human)}{
}}

\item{beta_j0_env}{Baseline environment-to-human transmission rate for each location. Must be a numeric vector of length equal to \code{location_id} and values greater than or equal to zero.}

\item{theta_j}{Proportion of the population with adequate Water, Sanitation, and Hygiene (WASH). Must be a numeric vector of length equal to \code{location_id} and values between 0 and 1.}

\item{psi_jt}{A matrix of environmental suitability values for V. cholerae. Must have rows equal to \code{length(location_id)} and columns equal to the daily sequence from \code{date_start} to \code{date_stop}, with values between 0 and 1.}

\item{zeta_1}{Shedding rate of V. cholerae into the environment by individuals with severe infection. Must be a numeric scalar greater than zero.}

\item{zeta_2}{Shedding rate of V. cholerae into the environment by individuals with mild or asymptomatic infection. Must be a numeric scalar greater than zero.}

\item{kappa}{Concentration of V. cholerae required for 50\% infection probability. Must be a numeric scalar greater than zero.}

\item{delta_min}{Minimum environmental decay rate of V. cholerae. Must be a numeric scalar greater than zero.}

\item{delta_max}{Maximum environmental decay rate of V. cholerae. Must be a numeric scalar greater than zero.}
}
\value{
Returns the validated list of parameters. If \code{output_file_path} is provided, the parameters are written to a JSON file.
}
\description{
This function generates a JSON file to be used as LASER model simulation parameters. The function validates all
input parameters to ensure they meet the required constraints for the dynamic transmission
model simulation.
}
\examples{
make_LASER_config(
     output_file_path = "parameters.json",
     date_start = "2024-12-01",
     date_stop = "2024-12-31",
     location_id = 1:2,
     location_name = c("Location A", "Location B"),
     N_j_initial = c("Location A" = 1000, "Location B" = 1000),
     S_j_initial = c("Location A" = 900, "Location B" = 900),
     E_j_initial = c("Location A" = 0, "Location B" = 0),
     I_j_initial = c("Location A" = 50, "Location B" = 50),
     R_j_initial = c("Location A" = 50, "Location B" = 50),
     V1_j_initial = c("Location A" = 0, "Location B" = 0),
     V2_j_initial = c("Location A" = 0, "Location B" = 0),
     b_jt = matrix(data = 0.0015, nrow = 2, ncol = 31),
     d_jt = matrix(data = 0.001, nrow = 2, ncol = 31),
     nu_jt = matrix(data = 0, nrow = 2, ncol = 31),
     phi_1 = 0.8,
     phi_2 = 0.85,
     omega_1 = 0.1,
     omega_2 = 0.12,
     epsilon = 0.05,
     gamma_1 = 0.2,
     gamma_2 = 0.25,
     mu = 0.01,
     rho = 0.9,
     sigma = 0.5,
     beta_j0_hum = c(0.05, 0.03),
     beta_j_seasonality = matrix(0, nrow = 2, ncol = 366),
     tau_i = c(0.1, 0.2),
     pi_ij = matrix(c(0.8, 0.2, 0.2, 0.8), nrow = 2),
     alpha_1 = 0.95,
     alpha_2 = 1,
     beta_j0_env = c(0.02, 0.04),
     theta_j = c(0.6, 0.7),
     psi_jt = matrix(data = 0, nrow = 2, ncol = 31),
     zeta_1 = 0.5,
     zeta_2 = 0.4,
     kappa = 10^5,
     delta_min = 0.1,
     delta_max = 0.01
)

}
