% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CFR_by_country.R
\name{plot_CFR_by_country}
\alias{plot_CFR_by_country}
\title{Plot Case Fatality Ratios, Total Cases, and Beta Distributions by Country}
\usage{
plot_CFR_by_country(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the cholera data and figures will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_WHO_ANNUAL}: Path to the directory where processed WHO annual cholera data is located.
\item \strong{DOCS_FIGURES}: Path to the directory where the plots will be saved.
}}
}
\value{
A list containing the two ggplot objects: one for CFR and total cases, and one for Beta distributions.
}
\description{
This function creates two separate plots: one for case fatality ratios (CFR) with confidence intervals and total cases by country, and another for Beta distributions for selected countries. The plots are generated based on cholera data stored in the \code{PATHS$DATA_WHO_ANNUAL} directory.
}
\examples{
\dontrun{
PATHS <- get_paths()
plot_CFR_by_country(PATHS)
}
}
