% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ENSO_data.R
\name{plot_ENSO_data}
\alias{plot_ENSO_data}
\title{Plot ENSO and DMI Data}
\usage{
plot_ENSO_data(PATHS, frequency)
}
\arguments{
\item{PATHS}{A list containing paths where ENSO and DMI data are stored and where output figures should be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_ENSO}: Path to the directory containing the compiled ENSO and DMI data.
\item \strong{DOCS_FIGURES}: Path to the directory where the output figure will be saved.
}}
}
\value{
A ggplot facet plot showing the time series of DMI, ENSO3, ENSO34, and ENSO4 over time. The plot includes:
\itemize{
\item A horizontal solid line at zero to indicate the neutral sea surface temperature anomaly.
\item A vertical dashed line to indicate the current date, with the current date displayed as vertical text.
\item Facets for each variable, allowing for easy comparison of DMI and ENSO regions.
}
}
\description{
This function creates a facet plot to visualize ENSO and DMI data over time, with separate panels for each variable (DMI, ENSO3, ENSO34, ENSO4). A horizontal line at zero is included to indicate the neutral value, and a vertical line marks the current date, with the date displayed vertically on the left side of the line.
}
\details{
This function uses \code{ggplot2} to create a line plot with facets for each variable (DMI, ENSO3, ENSO34, and ENSO4). It adds a horizontal reference line at zero and a vertical line representing the current date. The current date is displayed as vertical text aligned to the left of the vertical line. The data is expected to be in a \code{date}, \code{variable}, and \code{value} format.
}
\examples{
\dontrun{
# Load the ENSO and DMI data and plot it
PATHS <- get_paths()
plot_ENSO_data(PATHS)
}
}
