% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_plots.R
\name{plot_npe_diagnostics_coverage_bars}
\alias{plot_npe_diagnostics_coverage_bars}
\title{Plot NPE Coverage Diagnostic Bars}
\usage{
plot_npe_diagnostics_coverage_bars(
  diagnostics,
  plots_dir = NULL,
  max_params_per_page = 40,
  verbose = TRUE
)
}
\description{
Plot NPE Coverage Diagnostic Bars
}
