% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe_plots.R
\name{plot_npe_diagnostics_sbc}
\alias{plot_npe_diagnostics_sbc}
\title{Plot NPE diagnostic status table}
\usage{
plot_npe_diagnostics_sbc(
  diagnostics,
  plots_dir = NULL,
  max_params_per_page = 40,
  verbose = TRUE
)
}
\description{
Plot NPE diagnostic status table
}
