% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_transmission.R
\name{plot_seasonal_transmission}
\alias{plot_seasonal_transmission}
\title{Plot Precipitation and Cholera Case Data with Fourier Series Fits}
\usage{
plot_seasonal_transmission(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where processed climate, cholera, and fitted data are stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DOCS_FIGURES}: Path to save the output plot.
}}
}
\value{
The function generates and saves a PNG file showing the seasonal dynamics of precipitation and cholera cases across multiple countries.
}
\description{
This function visualizes the seasonal precipitation and cholera cases data, fitted with a double Fourier series model. It generates a plot of the Z-scored data for each country, indicating inferred countries where cholera cases data was imputed from neighbors.
}
