% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_transmission_example.R
\name{plot_seasonal_transmission_example}
\alias{plot_seasonal_transmission_example}
\title{Plot Seasonal Transmission Example with Grid Points}
\usage{
plot_seasonal_transmission_example(PATHS, country_iso_code, n_points = 30)
}
\arguments{
\item{PATHS}{A list containing paths where processed climate, cholera, and fitted data are stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_SHAPEFILES}: Path to the directory containing shapefiles for African countries.
\item \strong{DOCS_TABLES}: Path to load the processed precipitation and fitted values data.
\item \strong{DOCS_FIGURES}: Path to save the output plot.
}}

\item{country_iso_code}{A character string representing the ISO3 code of the country to plot.}

\item{n_points}{An integer specifying the number of grid points to generate within the country for sampling.}
}
\value{
The function generates and saves a PNG file showing the grid points on the country map and the seasonal transmission dynamics.
}
\description{
This function plots the map of a selected country with grid points used for climate data sampling, and visualizes the seasonal transmission dynamics using Fourier series for both precipitation and cholera cases.
}
