% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_suitability_and cases.R
\name{plot_suitability_and_cases}
\alias{plot_suitability_and_cases}
\title{Plot Predicted Suitability and Reported Cholera Cases for a Specific Country}
\usage{
plot_suitability_and_cases(PATHS, plot_iso_code)
}
\arguments{
\item{PATHS}{A list of paths to the directories where data is stored and plots will be saved.}

\item{plot_iso_code}{The ISO code of the country to plot.}
}
\value{
A combined plot of reported cholera cases and predicted environmental suitability.
}
\description{
This function plots the reported cholera cases and predicted environmental suitability
for a specific country over time, with shaded regions indicating cholera outbreaks.
}
