% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_suitability_by_country.R
\name{plot_suitability_by_country}
\alias{plot_suitability_by_country}
\title{Plot Predicted Suitability for All Countries}
\usage{
plot_suitability_by_country(PATHS)
}
\arguments{
\item{PATHS}{A list of paths to the directories where data is stored and plots will be saved.}
}
\value{
A facet plot of predicted environmental suitability for all countries.
}
\description{
This function generates a facet plot of predicted environmental suitability for all countries
over time, with shaded regions indicating cholera outbreaks.
}
