% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vaccine_effectiveness.R
\name{plot_vaccine_effectiveness}
\alias{plot_vaccine_effectiveness}
\title{Plot Vaccine Effectiveness and Beta Distribution}
\usage{
plot_vaccine_effectiveness(PATHS)
}
\arguments{
\item{PATHS}{A list containing the paths where the figures will be saved. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DOCS_FIGURES}: Path to the directory where the figures will be saved.
}}
}
\description{
This function generates two plots: one showing the vaccine effectiveness decay over time, and another showing the Beta distribution for the initial vaccine effectiveness.
}
