% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_initial_S.R
\name{print.mosaic_initial_conditions_S}
\alias{print.mosaic_initial_conditions_S}
\title{Print method for S compartment initial conditions}
\usage{
\method{print}{mosaic_initial_conditions_S}(x, ...)
}
\description{
Print method for S compartment initial conditions
}
