% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CFR_data.R
\name{process_CFR_data}
\alias{process_CFR_data}
\title{Process Cholera Data to Calculate Aggregated Case Fatality Ratios and Fit Beta Distributions (2014-2024)}
\usage{
process_CFR_data(PATHS, min_obs)
}
\arguments{
\item{PATHS}{A list containing the paths where the cholera data is located. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_WHO_ANNUAL}: Path to the directory containing processed WHO annual cholera data.
\item \strong{DOCS_TABLES}: Path to the directory where processed CFR data will be saved.
}}

\item{min_obs}{An integer specifying the minimum number of cases required for a country to have its own CFR. Countries with fewer cases than this threshold will have their CFR set to the value of the AFRO Region.}
}
\value{
A data frame with aggregated CFR, confidence intervals, and Beta distribution parameters for each country.
}
\description{
This function processes cholera data, aggregates cases and deaths for each country over the period 2014-2024, calculates the aggregated Case Fatality Ratio (CFR), and fits Beta distributions for each country. Countries with fewer cases than \code{min_obs} will have their CFR set to the value of the AFRO Region.
}
\examples{
\dontrun{
PATHS <- get_paths()
cholera_data <- process_CFR_data(PATHS, min_obs = 100)
}
}
