% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ENSO_data.R
\name{process_ENSO_data}
\alias{process_ENSO_data}
\title{Compile ENSO and DMI Data (Historical and Forecast)}
\usage{
process_ENSO_data(year_start = NULL, frequency = "monthly", method = "linear")
}
\arguments{
\item{year_start}{An integer representing the start year for filtering the data. Only data from this year onward will be included in the compiled data. The value must be greater than or equal to 1870.}

\item{frequency}{A character string specifying the time resolution of the output data. Valid options are "daily", "weekly", or "monthly".}

\item{method}{A character string specifying the interpolation method to use. Valid options are "linear" (for linear interpolation using \code{zoo::na.approx()}) or "spline" (for spline interpolation using \code{zoo::na.spline()}).}
}
\value{
A data frame with combined historical and forecast data for DMI, ENSO3, ENSO34, and ENSO4. The data frame includes the following columns:
\itemize{
\item \code{date}: The date of the data (daily, weekly, or monthly) in YYYY-MM-DD format.
\item \code{variable}: The variable name, which can be "DMI", "ENSO3", "ENSO34", or "ENSO4".
\item \code{value}: The value of the variable (sea surface temperature anomaly).
\item \code{year}: The year corresponding to the date.
\item \code{month}: The month corresponding to the date.
\item \code{month_name}: The name of the month (e.g., "Jan").
\item \code{week}: The week of the year (for "weekly" and "daily" frequency).
\item \code{doy}: The day of the year (only for "daily" frequency).
\item \code{date_start}: Start date for the week or month.
\item \code{date_stop}: End date for the week or month.
}
}
\description{
This function compiles historical and forecast data for DMI and ENSO (Niño3, Niño3.4, and Niño4) into a single data frame. The data is filtered to only include years from a specified \code{year_start} onwards. The function also allows for disaggregation of monthly data into daily or weekly values using either linear interpolation or spline interpolation.
}
\examples{
\dontrun{
# Compile the ENSO and DMI data from the year 2000 onwards
compiled_enso_data <- process_ENSO_data(2010, "monthly")

# Display the compiled data
head(compiled_enso_data)
}

}
