% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processs_suitability_data.R
\name{process_suitability_data}
\alias{process_suitability_data}
\title{Process Environmental Suitability Data for Cholera Transmission}
\usage{
process_suitability_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where the data is stored. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_CLIMATE}: Path to the directory where processed climate data is stored.
\item \strong{DATA_ENSO}: Path to the directory where ENSO data is stored.
\item \strong{DATA_WHO_WEEKLY}: Path to the directory containing processed weekly cholera cases data.
}}
}
\value{
This function processes the data and merges the climate, ENSO, and cholera cases data into a single dataset. The processed dataset is saved as a CSV file.
}
\description{
This function loads climate data, ENSO data, and weekly cholera cases data, processes them, and merges them into a dataset to estimate the environmental suitability for cholera transmission based on various environmental factors.
}
