% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_UN_demographics_data.R
\name{process_UN_demographics_data}
\alias{process_UN_demographics_data}
\title{Process UN World Prospects Data for African Countries}
\usage{
process_UN_demographics_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths where raw and processed data are stored.
PATHS is typically the output of the \code{get_paths()} function and should include the following components:
\itemize{
\item \strong{DATA_RAW}: Path to the directory containing the raw UN World Population Prospects data.
\item \strong{DATA_PROCESSED}: Path to the directory where processed demographic data should be saved.
}}
}
\value{
The function does not return a value. It processes the UN demographic data and saves the aggregated results as a CSV file in the specified directory.
}
\description{
This function processes demographic data from the UN World Population Prospects for African countries.
It loads available data, gets total population, birth rates, and death rates, and saves the processed data as a CSV file.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads raw demographic data from a CSV file.
\item Gets the population, birth rate, and death rate per day for each country.
\item Saves the processed data as a CSV file in the processed data directory.
}
The processed data file will be saved in the \verb{PATHS$DATA_PROCESSED/demographics/} directory, with a filename that includes the year range of the available data.

Data Source: \href{https://population.un.org/wpp/}{United Nations World Population Prospects}
}
\examples{
\dontrun{
# Define paths for raw and processed data using get_paths()
PATHS <- get_paths()

# Process the UN World Prospects data for all available years and save the results
process_demographics_data(PATHS)
}

}
