% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_WHO_vaccination_data.R
\name{process_WHO_vaccination_data}
\alias{process_WHO_vaccination_data}
\title{Process WHO Vaccination Data}
\usage{
process_WHO_vaccination_data(PATHS, max_rate_per_day)
}
\arguments{
\item{PATHS}{A list of file paths, which should include the following elements:
\describe{
\item{DATA_SCRAPE_WHO_VACCINATION}{The path to the folder containing the scraped WHO vaccination data.}
\item{DATA_DEMOGRAPHICS}{The path to the folder containing demographic data for population information.}
\item{MODEL_INPUT}{The path to the folder where the processed data and parameter files will be saved.}
}}

\item{max_rate_per_day}{The maximum vaccination rate per day for dose redistribution. Default is 100,000 doses/day.}
}
\value{
This function returns no R object, but it saves two CSV files:
\itemize{
\item A parameter data frame for the vaccination rate (nu) saved as \code{"param_nu_vaccination_rate.csv"}.
\item The full redistributed vaccination data saved as \code{"data_vaccinations.csv"}.
}
}
\description{
This function processes WHO vaccination request data, redistributes doses based on a maximum daily rate, and saves the processed data
into a format suitable for use in the MOSAIC cholera model. It also fills in missing campaign dates, infers decision dates when needed,
calculates cumulative doses, and computes vaccination proportions.
}
\details{
The function performs the following steps:
\itemize{
\item Loads WHO vaccination request data from 2016 to present and population data for 2023.
\item Converts country names to ISO codes and removes rows for countries not included in MOSAIC's ISO code list.
\item Infers missing campaign dates using the decision dates and calculates mean delays between decision and campaign dates.
\item Redistributes shipped doses day by day, based on the maximum daily rate.
\item Ensures that the full date range from the minimum distribution date to the current date is covered for all ISO codes, filling in any missing dates.
\item Calculates cumulative doses, the proportion of the population vaccinated, and other related metrics.
\item Merges population data and computes vaccination rates.
\item Saves the processed vaccination data and parameter data frame for use in the MOSAIC cholera model.
}
}
\examples{
\dontrun{
PATHS <- list(
  DATA_SCRAPE_WHO_VACCINATION = "path/to/who_vaccination_data",
  DATA_DEMOGRAPHICS = "path/to/demographics",
  MODEL_INPUT = "path/to/save/processed/data"
)
process_WHO_vaccination_data(PATHS, max_rate_per_day = 100000)
}

}
