% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_WHO_weekly_data.R
\name{process_WHO_weekly_data}
\alias{process_WHO_weekly_data}
\title{Process Weekly Cholera Data: Convert Country Names to ISO3 and Filter by AFRO Countries}
\usage{
process_WHO_weekly_data(PATHS)
}
\arguments{
\item{PATHS}{A list containing paths to the raw and processed data directories. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_RAW}: Path to the raw cholera data file (e.g., \code{cholera_country_weekly.csv}).
\item \strong{DATA_PROCESSED}: Path to save the processed cholera data (e.g., \code{cholera_country_weekly_processed.csv}).
}}
}
\value{
The function does not return a value but processes the cholera data and saves it to a specified file as a CSV.
}
\description{
This function processes cholera data by converting country names to ISO3 codes, filtering for AFRO region countries, and cleaning and organizing the data for further analysis.
}
\details{
The function reads in cholera data, converts country names to ISO3 codes (and vice versa), filters for countries in the AFRO region, and ensures the data is cleaned and organized by year and week. It also handles special cases such as renaming the Democratic Republic of Congo and Congo (Republic) and modifies column names for consistency.
}
\examples{
\dontrun{
# Assuming PATHS is generated from get_paths()
PATHS <- get_paths()

# Process cholera data
process_WHO_weekly_data(PATHS)
}

}
