% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_WHO_weekly_data.R
\name{process_WHO_weekly_data}
\alias{process_WHO_weekly_data}
\title{Process Weekly Cholera Data: Convert Country Names to ISO3 and Filter by WHO AFRO Region with Cutoff for Binary Cases}
\usage{
process_WHO_weekly_data(PATHS, cutoff)
}
\arguments{
\item{PATHS}{A list containing paths to the raw and processed data directories. Typically generated by the \code{get_paths()} function and should include:
\itemize{
\item \strong{DATA_SCRAPE_WHO_WEEKLY}: Path to the raw cholera data file (e.g., \code{cholera_country_weekly.csv}).
\item \strong{DATA_WHO_WEEKLY}: Path to save the processed cholera data (e.g., \code{cholera_country_weekly_processed.csv}).
}}

\item{cutoff}{An integer specifying the case count threshold above which a binary outbreak indicator is set to 1. Specific countries (e.g., Cameroon) can have adjusted cutoff values.}
}
\value{
The function processes cholera data, adds a binary outbreak indicator, and saves the result as a CSV file. No value is returned.
}
\description{
This function processes weekly cholera data by converting country names to ISO3 codes, filtering the data for countries in the WHO AFRO region, and organizing the data by year and week. It adds binary outbreak indicators based on a specified case cutoff value and generates start and stop dates for each ISO week.
}
\details{
This function performs the following tasks:
\itemize{
\item Reads raw cholera data from a CSV file.
\item Converts country names to ISO3 codes using the \code{MOSAIC::convert_country_to_iso} and \code{MOSAIC::convert_iso_to_country} functions.
\item Filters the dataset to include only countries in the WHO AFRO region (based on \code{MOSAIC::iso_codes_who_afro}).
\item Orders the data by year, week, and country.
\item Renames columns for consistency: \code{cases_by_week} to \code{cases} and \code{deaths_by_week} to \code{deaths}.
\item Adds \code{date_start} and \code{date_stop} columns based on ISO week, using the \code{ISOweek::ISOweek2date} function.
\item Adds a binary outbreak indicator (\code{cases_binary}) based on a specified \code{cutoff} for case counts, with specific adjustments for certain countries (e.g., Cameroon).
\item Reorganizes the columns for clarity and saves the processed data to the specified path.
}
}
\examples{
\dontrun{
# Define paths for raw and processed cholera data
PATHS <- get_paths()

# Process the weekly cholera data with a cutoff of 700 cases
process_WHO_weekly_data(PATHS, cutoff = 700)
}
}
