% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_statistics.R
\name{weighted_quantiles}
\alias{weighted_quantiles}
\title{Weighted quantiles}
\usage{
weighted_quantiles(x, w, probs)
}
\arguments{
\item{x}{Numeric vector of values}

\item{w}{Numeric vector of weights (same length as x)}

\item{probs}{Numeric vector of quantile probabilities (between 0 and 1)}
}
\value{
Vector of weighted quantiles
}
\description{
Calculates weighted quantiles for a vector of values using linear interpolation.
}
\details{
Sorts values and weights, calculates cumulative weight distribution,
and uses linear interpolation to estimate quantiles at specified probability levels.
}
\examples{
x <- c(1, 2, 3, 4, 5)
w <- c(0.1, 0.2, 0.4, 0.2, 0.1)
weighted_quantiles(x, w, c(0.25, 0.5, 0.75))

}
